+++ lib/src/widget/auto_ble_dfu.dart
@@ -7,6 +7,8 @@ class AutoBleDfu extends StatelessObserverWidget {
   const AutoBleDfu({super.key});
 
   Future<void> _autoDfu() async {
+    if(NrfBleDfu().file.datPath == null) return;
+    if(NrfBleDfu().file.binPath == null) return;
     if (!FlutterBluePlus.isScanningNow) await FlutterBluePlus.startScan();
     if (!NrfBleDfu().setup.enableTargetEntryProcess) NrfBleDfu().setup.autoDfuTargets.clear();
     final scanResults = await FlutterBluePlus.scanResults.first;