+++ CHANGELOG.md
@@ -1,8 +1,11 @@
+## 0.0.4
+* Add button to mark completed devices.
+
 ## 0.0.3
-* Does not auto DFU for finished devices 
+* Does not auto DFU for finished devices.
 
 ## 0.0.2
-* Add automated dfu mode entry
+* Add automated dfu mode entry.
 
 ## 0.0.1
 * Initial release.